requirements.txt
asgiref==3.5.2
attrs==22.1.0
Django==4.1.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
drf-spectacular==0.24.2
drf-spectacular-sidecar==2022.9.1
flake8==5.0.4
gunicorn==20.1.0
inflection==0.5.1
isort==5.10.1
jsonschema==3.2.0
Markdown==3.4.1
mccabe==0.7.0
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.5.0
pyrsistent==0.18.1
pytz==2022.2.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
